requirements.txt
Django==1.9.13
easy-thumbnails==2.4.1
django-ckeditor==5.0.3
django-cleanup==1.0.1
django-crequest==2016.3.16
django-jet==1.0.6
django-modeltranslation==0.11  # 0.12.1
django-multiselectfield==0.1.8
django-request==1.5.3
django-rosetta==0.7.13
ipython==2.0.0
ipdb==0.10.3
solid_i18n==1.4.1
# uWSGI
# psycopg2
